requirements.txt
appnope==0.1.0

astroid==2.2.5
attrs==19.1.0
autopep8==1.4.4
backcall==0.1.0
bleach==3.1.0
boltons==19.2.0
Click==7.0
couchbase==2.5.8
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
entrypoints==0.3
Flask==1.1.1
gcloud==0.18.3
googleapis-common-protos==1.6.0
httplib2==0.14.0
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.1
jsonschema==3.0.2

kiwisolver==1.1.0
lazy-object-proxy==1.4.2


mccabe==0.6.1
mistune==0.8.4
mpmath==1.1.0
mypy-extensions==0.4.3
nbconvert==5.6.0
nbformat==4.4.0
nose==1.3.7
notebook==6.0.1
numpy==1.17.2
oauth2client==4.1.3

pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.10.0
ptyprocess==0.6.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycodestyle==2.5.0
Pygments==2.4.2
pylint==2.3.1
pyparsing==2.4.2
requests==2.2.0
pyrsistent==0.15.4

python-dateutil==2.8.0
pytz==2019.2
pyzmq==18.1.0
qtconsole==4.5.5


Send2Trash==1.5.0
six==1.12.0
sympy==1.4
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
traitlets==4.3.2
typed-ast==1.4.0
tzlocal==2.0.0
urllib3==1.25.6
wcwidth==0.1.7
webencodings==0.5.1

widgetsnbextension==3.5.1
wrapt==1.11.2
